src/mcp/requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
PyMuPDF==1.24.0
